Unified the computation of the required memory that needs to be allocated when loading a nifti-file.

+++ src/rsbandpass_common.h
@@ -27,8 +27,8 @@ struct rsBandpassParameters {
     short zDim;
     short vDim;
     
-	short pixtype;
-	size_t dt;
+	size_t pixtype;
+	short dt;
     
     float inter;
     float slope;
