+++ src/rsregression_common.h
@@ -33,8 +33,8 @@ struct rsRegressionParameters {
     short zDim;
     short vDim;
     
-	short pixtype;
-	size_t dt;
+	size_t pixtype;
+	short dt;
     
     float inter;
     float slope;
